+++ dandy/llm/bot/llm_bot.py
@@ -1,6 +1,6 @@
 from abc import ABC
 
-from typing_extensions import Type, Generic
+from typing_extensions import Type, Generic, Union
 
 from dandy.bot import BaseBot
 from dandy.bot.exceptions import BotException
@@ -16,7 +16,7 @@ from dandy.llm.service.config.options import LlmConfigOptions
 class BaseLlmBot(BaseBot, ABC, Generic[IntelType]):
     config: str = 'DEFAULT'
     config_options: LlmConfigOptions = LlmConfigOptions()
-    instructions_prompt: Prompt
+    instructions_prompt: Prompt = Prompt("You're a helpful assistant please follow the users instructions.")
     intel_class: Type[BaseIntel] = DefaultLlmIntel
 
     def __new__(cls):
@@ -30,19 +30,23 @@ class BaseLlmBot(BaseBot, ABC, Generic[IntelType]):
     @classmethod
     def process_prompt_to_intel(
             cls,
-            prompt: Prompt,
+            prompt: Union[Prompt, str],
             intel_class: Type[IntelType],
+            postfix_system_prompt: Union[Prompt, None] = None
     ) -> IntelType:
+        
+        prefix_system_prompt = Prompt()
+        prefix_system_prompt.prompt(cls.instructions_prompt)
+        
+        if postfix_system_prompt:
+            prefix_system_prompt.prompt(postfix_system_prompt)
 
         return llm_configs[cls.config].generate_service(
             llm_options=cls.config_options
         ).process_prompt_to_intel(
-            prompt=prompt,
+            prompt=prompt if isinstance(prompt, Prompt) else Prompt(prompt),
             intel_class=intel_class,
-            prefix_system_prompt=(
-                Prompt()
-                .prompt(cls.instructions_prompt)
-            )
+            prefix_system_prompt=prefix_system_prompt
         )
 
     @classmethod
@@ -51,8 +55,6 @@ class BaseLlmBot(BaseBot, ABC, Generic[IntelType]):
 
 
 class LlmBot(BaseLlmBot, Generic[IntelType]):
-    instructions_prompt: Prompt = Prompt("You're a helpful assistant.")
-
     @classmethod
     def process(
             cls,